+++ test/loader.test.js
@@ -1,62 +1,64 @@
 'use strict';
 
+const fetch = require('node-fetch');
+const loader = require('../lib/loader.js');
+const nock = require('nock');
 const path = require('path');
 const yaml = require('js-yaml');
-const loader = require('../lib/loader.js');
 
 describe('loader.js', () => {
-    describe('loadRules()', () => {
-        const expectedRules = {
-            default: [
-                "parameter-description",
-                "operation-operationId",
-                "operation-summary-or-description",
-                "operation-tags",
-                "server-trailing-slash",
-                "openapi-tags",
-                "openapi-tags-alphabetical",
-                "reference-no-other-properties",
-                "example-value-or-externalValue",
-                // "reference-components-regex",
-                "no-script-tags-in-markdown",
-                "info-contact",
-                "license-apimatic-bug",
-                "no-eval-in-descriptions",
-            ],
-            strict: [
-                "contact-properties",
-                "license-url",
-                "server-not-example.com",
-                "tag-description",
-            ],
-            wework: [
-                "x-tagGroups",
-                "short-summary",
-            ]
-        };
-
+    describe('loadRuleFiles()', () => {
         it('load default rules', () => {
-            const loadedNames = loader.loadRules(['default']).map(x => x.name)
-            should(loadedNames).be.eql(
-                expectedRules.default
-            );
+            loader.loadRuleFiles(['default']).should.be.fulfilledWith(['default']);
         });
 
-        it('load default rules', () => {
-            const loadedNames = loader.loadRules(['strict']).map(x => x.name)
-            should(loadedNames).be.eql(
-                expectedRules['default'].concat(expectedRules['strict'])
-            );
+        it('load strict rules', () => {
+            loader.loadRuleFiles(['strict']).should.be.fulfilledWith(['strict', 'default']);
         });
 
-        it('load default rules', () => {
-            const loadedNames = loader.loadRules(['wework']).map(x => x.name)
-            should(loadedNames).be.eql(
-                expectedRules['default'].concat(
-                    expectedRules['strict'],
-                    expectedRules['wework']
-                )
-            );
+
+        it('load default & strict rules', () => {
+            loader.loadRuleFiles(['strict', 'default']).should.be.fulfilledWith(['strict', 'default']);
+        });
+
+        it('load wework rules', () => {
+            loader.loadRuleFiles(['wework']).should.be.fulfilledWith(['wework', 'strict', 'default']);
+        });
+
+        context('when loading from url', () => {
+            const host = 'http://example.org';
+            const url = host + '/';
+
+            it('retrieves rules from valid url', () => {
+                nock(host).get('/').replyWithFile(200, __dirname + '/../rules/strict.json', {
+                    'Content-Type': 'application/json'
+                });
+
+                loader.loadRuleFiles([host + '/']).should.be.fulfilledWith([host + '/', 'default']);
+            });
+
+            context('when the file being loaded is garbage', () => {
+                const setupMock = () => nock(host).get('/').reply(200, 'this is not json AT ALL');
+
+                it('reject with a ReadError', () => {
+                    setupMock();
+                    loader.loadRuleFiles([url]).should.be.rejectedWith(loader.ReadError);
+                });
+
+                it('error should be useful', () => {
+                    setupMock();
+                    loader.loadRuleFiles([url]).should.be.rejectedWith(/Invalid JSON: invalid json response body/);
+                });
+            })
+
+            context('when the url is not found', () => {
+                const setupMock = () => nock(host).get('/').reply(404);
+
+                it('rejects with an OpenError', () => {
+                    setupMock();
+                    loader.loadRuleFiles([url]).should.be.rejectedWith(loader.OpenError);
+                });
+            });
         });
     });
 
@@ -114,15 +116,8 @@ describe('loader.js', () => {
             should(thrownError.name).equal('OpenError');
         });
 
-        it('throws ReadError for invalid YAML/JSON', async () => {
-            let thrownError;
-            try {
-                await loader.loadSpec(samplesDir + 'not-openapi.txt');
-            }
-            catch (error) {
-                thrownError = error;
-            }
-            should(thrownError.name).equal('ReadError');
+        it('throws ReadError for invalid YAML/JSON', () => {
+            loader.loadSpec(samplesDir + 'not-openapi.txt').should.be.rejectedWith(loader.ReadError);
         });
     });
 });